requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dill==0.3.2
Django==3.0.8
django-cors-headers==3.4.0
django-password-strength==1.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
EasyProcess==0.3
geckodriver-autoinstaller==0.1.0
idna==2.10
lxml==4.5.2
MouseInfo==0.1.3
password-strength==0.0.3.post2
Pillow==7.2.0
psycopg2-binary==2.8.5
PyAutoGUI==0.9.50
PyAutoIt==0.6.0
PyGetWindow==0.0.8
PyJWT==1.7.1
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
